SwiftBrowser errantly builds with iOS SDK on watchOS and tvOS. https://bugs.webkit.org/show_bug.cgi?id=300968 rdar://162843312
Reviewed by Abrar Rahman Protyasha.

When building using the watchos/watchsimulator or appletvos/appletvsimulator
SDKs, we don't want to build SwiftBrowser. To achieve this, we removed those
from `SUPPORTED_PLATFORMS`; however, this had the unintended side effect of
causing the build system to fallback to using the iphoneos/iphonesimulator SDK
to build SwiftBrowser.

So, this PR adds `appletvos appletvsimulator watchos watchsimulator ` to
`SUPPORTED_PLATFORMS` in the relevant .xcconfig file (listed below), and has a
couple of drive-by fixes:
- Fix mismatched parens on line 47.
- Allow SwiftBrowser to build on visionOS 26 (fix SDK asterisk syntax error).

* Tools/SwiftBrowser/Configurations/Base.xcconfig: see above.

Canonical link: https://commits.webkit.org/301715@main