[WebGPU Swift] Enable runtime flag by default when compiler is compatible https://bugs.webkit.org/show_bug.cgi?id=300935 rdar://162791460
Reviewed by Mike Wyrzykowski.

Change the runtime flag that enables WebGPU Swift such that if the compiler is new
enough to support all the features we need, then it's enabled by default. It remains
possible to override it and turn it on off on any compiler version.

* Source/WebGPU/WebGPU.xcodeproj/project.pbxproj:
* Source/WebGPU/WebGPU/Capabilities.swift: Added.
(swiftCompilerSupportsWebGPU):
* Source/WebGPU/WebGPU/HardwareCapabilities.mm:
(WebGPU::isWebGPUSwiftEnabled):

Canonical link: https://commits.webkit.org/301749@main