Find in Page doesn’t correctly highlight results after revealing content on iOS/iPadOS https://bugs.webkit.org/show_bug.cgi?id=300744 rdar://159801954
Reviewed by Aditya Keerthi.

The issue is that the `FindDecorationStyle::Found` markers are issued before the content was revealed in the `FindDecorationStyle::Highlighted` branch of `decorateTextRangeWithStyle`.

This leads to the found markers failing to be added.

Since we can't change the order for which UIKit issues these markers, we instead:
1. store the markers that fail to be added into a set
2. re-attempt to add those markers if some content got expanded

Not testable, highlights are hard to test since they are cutouts in a layer and there is no existing testing infrastructure for them.

* Source/WebCore/dom/DocumentMarkerController.cpp:
(WebCore::DocumentMarkerController::addMarker):
* Source/WebCore/dom/DocumentMarkerController.h:
* Source/WebCore/dom/FindRevealAlgorithms.cpp:
(WebCore::revealClosedDetailsAndHiddenUntilFoundAncestors):
* Source/WebCore/dom/FindRevealAlgorithms.h:
* Source/WebKit/WebProcess/WebPage/WebFoundTextRangeController.cpp:
(WebKit::WebFoundTextRangeController::decorateTextRangeWithStyle):
(WebKit::WebFoundTextRangeController::clearAllDecoratedFoundText):
* Source/WebKit/WebProcess/WebPage/WebFoundTextRangeController.h:

Canonical link: https://commits.webkit.org/301751@main