[Tools][CMake] Avoid ccache misses when building in different directories https://bugs.webkit.org/show_bug.cgi?id=301003
Reviewed by Philippe Normand and Adrian Perez de Castro.

ccache by default stores the full path of the files as part of the hash,
so it is not possible to share the cache when building the same project
on different paths.

Fortunately, this can be tuned so ccache stores only relative paths for
the WebKit source codes, and this allows to hit the cache even when the
path to the WebKit source code changes.

This patch configures ccache for that automatically, so any developer
using the script build-webkit can reuse the same ccache results even
when it moves or renames the WebKit directory.

* Source/cmake/OptionsCommon.cmake:
* Tools/Scripts/build-webkit:

Canonical link: https://commits.webkit.org/301772@main