REGRESSION(300810@main): Caused media/media-webm-no-duration.html to crash https://bugs.webkit.org/show_bug.cgi?id=300940 rdar://162815321
Reviewed by Anne van Kesteren.

The test was crashing as it disabled `UseGPUProcessForMediaEnabled` preference
and it was required for the AudioVideoRendererRemote IPC service to be enabled.

The service didn't actually require UseGPUProcessForMediaEnabled.
The original test failed as the file being tested (media/content/test-no-duration.webm)
didn't even exists.
The original test was checking that a webm generated by Chrome's MediaRecorder
played. We create such file and add it to the test.

Update expectations with the correct file path (which was wrong too)

Re-enabled test and fix the original error that caused that test to always fails.
* LayoutTests/TestExpectations:
* LayoutTests/media/content/test-no-duration.webm: Added.
* LayoutTests/media/media-webm-no-duration-expected.txt:
* LayoutTests/media/media-webm-no-duration.html:
* LayoutTests/platform/mac/TestExpectations:
* Source/WebKit/GPUProcess/media/RemoteAudioVideoRendererProxyManager.messages.in:

Canonical link: https://commits.webkit.org/301775@main