[GTK][WPE] Offer a BUILD_WEBKIT_PRE_SCRIPT environment variable, to execute a script prior to build-webkit https://bugs.webkit.org/show_bug.cgi?id=301080
Reviewed by Carlos Alberto Lopez Perez.

If BUILD_WEBKIT_PRE_SCRIPT is set, verify it's pointing to an executable
script or complain. If it's unset, no actions are taken.

If it points to a valid executable, execute it before starting the
WebKit build. This is useful for the Gtk/WPE and the new SDK to support
remote ccache backends, where you need to execute a port knocking
procedure to activate the access to the ccache backend on the server.

* Tools/Scripts/build-webkit:

Canonical link: https://commits.webkit.org/301802@main