WKWebExtensionAPIMenus.MacEditableContextMenuItems is a constant failure https://bugs.webkit.org/show_bug.cgi?id=301106 rdar://163052124
Reviewed by Brian Weinstein.

The text here is being wrapped so when we go to right click on the text, the selectionText is "Text"
and not "Area". Prevent the text area from being wrapped so that the when we right click, the text
matches the expected value. Also, reduce the font size since it's not necessary for it to be 100px.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtensionAPIMenus.mm:
(TestWebKitAPI::TEST(WKWebExtensionAPIMenus, MacEditableContextMenuItems)):

Canonical link: https://commits.webkit.org/301825@main