Make TestWebKitAPI.WebTransport.ServerBidirectional stop timing out on bots https://bugs.webkit.org/show_bug.cgi?id=301192 rdar://163122624
Unreviewed.

Closing the writable stream before waiting for the readable stream causes the test
to time out because closing only one direction of a WebTransport stream but not the
other hasn't been implemented yet.  When it is implemented, we'll test it properly.
Until then, make the test test what it is designed to, that bidirectional streams
initiated by the server are hooked up correctly.  Do this without timing out.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/WebTransport.mm:
(TestWebKitAPI::TEST(WebTransport, ServerBidirectional)):

Canonical link: https://commits.webkit.org/301883@main