Adds Quirk so that Amazon Prime Video Safari users can watch on the Web https://bugs.webkit.org/show_bug.cgi?id=301171 rdar://117771731
Reviewed by Brent Fulgham.

Amazon Prime Video on iPad proposes users with a splash screen pushing
them to download the app without the possibility to watch on the Web.
This is not the case on similar Android Tablet with Chrome and a desktop
User Agent where users can continue to watch their favorite show on the
Web. This quirks makes it possible for Safari users to access the same
user experience and still leaving users with the choice to download the
app if they wish so. An outreach effort has been started with Amazon to
find a path which benefits both users of Safari and Amazon Prime Video.

* Source/WebCore/page/Quirks.cpp:
(WebCore::Quirks::needsCustomUserAgentOverride):
(WebCore::handleAmazonQuirks):

Canonical link: https://commits.webkit.org/301891@main