[visionOS] Entering docked mode from element fullscreen has brief spatial audio hiccup rdar://160143659 https://bugs.webkit.org/show_bug.cgi?id=301218
Reviewed by Eric Carlson.

It's possible for the AudioVideoRenderer to get a video target before it receives
a spatial tracking label. In this scenario, the SpatialExperienceHelper will create
a tethered experience with no label, confusing the audio rendering system.

Only create a tethered experience when all of the criteria are met, including that
the label is non-empty.

Drive-by fix: Enable logging for the spatial experience methods so that this behavior
can be seen in logs.

* Source/WebCore/platform/audio/cocoa/SpatialAudioExperienceHelper.mm:
(WebCore::createSpatialAudioExperienceWithOptions):
* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.mm:
(WebCore::AudioVideoRendererAVFObjC::setSpatialTrackingInfo):
(WebCore::AudioVideoRendererAVFObjC::updateSpatialTrackingLabel):

Canonical link: https://commits.webkit.org/301919@main