[MSE/SourceBuffer] MediaSourcePrivate doesn't need to know about CDMInstance/CDMSession https://bugs.webkit.org/show_bug.cgi?id=301073 rdar://163005164
Reviewed by Eric Carlson.

The MediaPlayerPrivateMediaSourceAVFObjC had to communicate to all the
SourceBufferPrivateAVFObjC to notify them if a new CDMInstance or LegacyCDMSession
was attached to the player.
Now that the AudioVideoRenderer handles the decryption of MediaSample
the MediaSourcePrivateAVFObjC and all attached SourceBufferPrivateAVFObjCs
do not need to be notified that a CDM has been attached. This knowledge
is the responsibility of the MediaPlayerPrivate and its AudioVideoRenderer.
We can remove all associated code.

No change in observable behaviour, covered by existing tests.

* Source/WebCore/platform/graphics/avfoundation/AudioVideoRendererAVFObjC.mm:
(WebCore::AudioVideoRendererAVFObjC::enqueueSample):
(WebCore::AudioVideoRendererAVFObjC::canEnqueueSample): Fix from 301391@main
didn't carry over in 301865@main. Causing seek with encrypted audio to never complete.
(WebCore::AudioVideoRendererAVFObjC::attachContentKeyToSampleIfNeeded):

Canonical link: https://commits.webkit.org/301921@main