[MiniBrowser] Fix logging channel names being normalized as URLs https://bugs.webkit.org/show_bug.cgi?id=301227 rdar://163146056
Reviewed by Abrar Rahman Protyasha and Basuke Suzuki.

After commits.webkit.org/301651@main, we treat strings without a scheme as a URL, and
normalize it with the proper scheme before passing it to MIniBrowser.

This caused issues with WebKit logging channels, which do not have schemes, and thus
were being treated as URLs and getting schemes added.

Fix this by only normalizing if it actually looks like a URL (to not accidentally normalize a logging stream):
         - starts with localhost
         - looks like an IP address
         - contains a dot (likely a domain) or slash (path)

* Tools/Scripts/webkitpy/minibrowser/run_webkit_app.py:
(main):

Canonical link: https://commits.webkit.org/301927@main