[WebXR] Implement XRWebGLBinding constructor https://bugs.webkit.org/show_bug.cgi?id=301102
Reviewed by Dan Glastonbury.

Adding the implementation of the constructor of the bindings
object as per the WebXR spec. So far is basically a sequence
of checks.

Replaced subtests expectations from FAIL to PASS and unskipping
the WPT test that verifies the constructor behaviour.

* LayoutTests/imported/w3c/web-platform-tests/webxr/layers/idlharness.https.window-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/webxr/layers/xrWebGLBinding_constructor.https-expected.txt:
* LayoutTests/platform/glib/TestExpectations:
* Source/WTF/Scripts/Preferences/UnifiedWebPreferences.yaml:
* Source/WebCore/Modules/webxr/XRWebGLBinding.cpp:
(WebCore::XRWebGLBinding::create):
(WebCore::XRWebGLBinding::XRWebGLBinding):
* Source/WebCore/Modules/webxr/XRWebGLBinding.h:

Canonical link: https://commits.webkit.org/301937@main