Add platform support to update-safer-cpp-expectations https://bugs.webkit.org/show_bug.cgi?id=301265
Reviewed by Chris Dumez.

Added the support for platform modifier in update-safer-cpp-expectations.

The script now supports --platform option to specify the platform for which
safer C++ expectations need to be updated. When the platform is omitted,
the addition or removal happens for both platforms (Mac and iOS).

When there is a Mac expectation, removal of the same file for iOS does not
have any effect (results in a warning emitted to stdout).

In addition, this PR adds a support for preserving comments in safer C++
expectation files. Each comment is assumed to belong to the file beneath it
for sorting purposes.

For now, all these changes are made for updating individual checker by
enumerating file names, not when -r / --unexpected-results-file is used.

* Tools/Scripts/update-safer-cpp-expectations:
(parser):
(update_expectations_from_file):
(modify_expectations_for_checker):
(update_expectations):
(main):

Canonical link: https://commits.webkit.org/301953@main