Wasm JS string builtins should be re-exported as WebAssemblyWrapperFunctions https://bugs.webkit.org/show_bug.cgi?id=301229 rdar://163013276
Reviewed by Yusuke Suzuki.

When a Wasm JS string builtin is re-exported by a module, we use the jsWrapper() of the
builtin (a JSFunction) as the exported object. This works for as long as the exported
function is called from JavaScript or imported into other modules via the normal import
route. However, the fuzzer case included as the test attempts to store that JSFunction
directly into a Wasm function table. That fails because a function table expects its
values to inherit from WebAssemblyFunctionBase, and a plain JSFunction doesn't. This
causes an assertion failure in debug and asan builds and a crash in release builds.

The patch tweaks export logic so that builtin jsWrapper() is wrapped into a
WebAssemblyWrapperFunction, just like re-exported "normal" JS functions.

Canonical link: https://commits.webkit.org/301963@main